+++ dev-packages/e2e-tests/test-applications/react-router-7-framework/app/routes/performance/server-action.tsx
@@ -1,0 +1,24 @@
+import { Form } from 'react-router';
+import type { Route } from './+types/server-action';
+
+export async function action({ request }: Route.ActionArgs) {
+  let formData = await request.formData();
+  let name = formData.get('name');
+  await new Promise(resolve => setTimeout(resolve, 1000));
+  return {
+    greeting: `Hola ${name}`,
+  };
+}
+
+export default function Project({ actionData }: Route.ComponentProps) {
+  return (
+    <div>
+      <h1>Server action page</h1>
+      <Form method="post">
+        <input type="text" name="name" />
+        <button type="submit">Submit</button>
+      </Form>
+      {actionData ? <p>{actionData.greeting}</p> : null}
+    </div>
+  );
+}