+++ packages/react-router/test/server/sdk.test.ts
@@ -3,7 +3,7 @@ import type { NodeClient } from '@sentry/node';
 import * as SentryNode from '@sentry/node';
 import { SDK_VERSION } from '@sentry/node';
 import { afterEach, describe, expect, it, vi } from 'vitest';
-import * as LowQualityModule from '../../src/server/lowQualityTransactionsFilterIntegration';
+import * as LowQualityModule from '../../src/server/integration/lowQualityTransactionsFilterIntegration';
 import { init as reactRouterInit } from '../../src/server/sdk';
 
 const nodeInit = vi.spyOn(SentryNode, 'init');